make gopkgs go mod v2 compliant
