+++ cmd/gopkgs/main.go
@@ -10,7 +10,7 @@ import (
 	"runtime/trace"
 	"text/tabwriter"
 
-	"github.com/uudashr/gopkgs"
+	"github.com/uudashr/gopkgs/v2"
 )
 
 var usageInfo = `
