+++ doc.go
@@ -1,2 +1,2 @@
 // Package gopkgs is a utility to get list of golang packages.
-package gopkgs
+package gopkgs // import "github.com/uudashr/gopkgs/v2"
