+++ go.mod
@@ -1,7 +1,8 @@
-module github.com/uudashr/gopkgs
+module github.com/uudashr/gopkgs/v2
 
 go 1.12
 
-require github.com/uudashr/gopkgs/v2 v2.1.2
-
-replace github.com/uudashr/gopkgs/v2 => ./v2
+require (
+	github.com/karrick/godirwalk v1.15.3
+	github.com/pkg/errors v0.9.1
+)
