+++ internal/doc.go
@@ -1,0 +1,2 @@
+// Package internal is a utility to get list of golang packages.
+package internal // import "github.com/uudashr/gopkgs/v2/internal"
