+++ v2/go.mod
@@ -1,8 +1,8 @@
 module github.com/uudashr/gopkgs/v2
 
-go 1.13
+go 1.12
 
 require (
-	github.com/karrick/godirwalk v1.12.0
-	github.com/pkg/errors v0.8.1
+	github.com/karrick/godirwalk v1.15.3
+	github.com/pkg/errors v0.9.1
 )
