+++ v2/internal/vendor.go
@@ -1,0 +1,7 @@
+package internal // import "github.com/uudashr/gopkgs/v2/internal"
+
+import "strings"
+
+func visibleVendor(workDir, vendorDir string) bool {
+	return strings.Index(workDir, vendorDir) == 0
+}
